# frontend/openapi/modular-pipeline-openapi.json
{
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "ai_cost_profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "balanced | aggressive | quality",
            "title": "Ai Cost Profile"
          },
          "context": {
            "description": "Document text chunks as context",
            "items": {
              "type": "string"
            },
            "title": "Context",
            "type": "array"
          },
          "history": {
            "description": "Previous conversation turns",
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "title": "History",
            "type": "array"
          },
          "llm": {
            "$ref": "#/components/schemas/LLMConfig"
          },
          "query": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "sampling": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "integer"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-task sampling overrides",
            "title": "Sampling"
          }
        },
        "required": [
          "query",
          "llm"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
...
      "GenerateSparksResponse": {
        "properties": {
          "sparks": {
            "items": {
              "$ref": "#/components/schemas/SparkResponse"
            },
            "title": "Sparks",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "sparks",
          "total"
        ],
        "title": "GenerateSparksResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "学术研究智能体 — 论文分析、知识管理与智能写作辅助平台",
    "title": "Scholar AI API",
    "version": "1.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/chat/ask": {
      "post": {
        "description": "Send user query + context to configured LLM and return the answer.",
        "operationId": "post_chat_ask",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat Ask",
        "tags": [
          "Chat"
        ]
      }
    },
    "/chat/models": {
      "get": {
        "description": "List known LLM model options. Optionally filter by provider.",
        "operationId": "get_chat_models",
        "parameters": [
          {
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelInfo"
                  },
                  "title": "Response Get Chat Models",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Supported Models",
        "tags": [
          "Chat"
        ]
      }
    },
...
    "/chat/stream": {
      "post": {
        "description": "Stream LLM response via Server-Sent Events (SSE).\n\nEach event is a JSON line:\n  data: {\"event\":\"text_delta\",\"delta\":\"...\"}\n  data: {\"event\":\"usage\",\"usage\":{...},\"model\":\"...\"}\n  data: {\"event\":\"done\"}\n  data: {\"event\":\"error\",\"error\":\"...\"}",
        "operationId": "post_chat_stream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatStreamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat Stream",
        "tags": [
          "Chat"
        ]
      }
    },
...
    "/inspiration/generate": {
      "post": {
        "description": "基于查询生成启发点列表。",
        "operationId": "post_inspiration_generate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateSparksRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateSparksResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Inspirations",
        "tags": [
          "Inspiration"
        ]
      }
    },
...
    "/memory/search": {
      "post": {
        "description": "Search MemPalace for project memory.",
        "operationId": "post_memory_search",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemorySearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemorySearchResponsePayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Memory",
        "tags": [
          "Memory"
        ]
      }
    },
...
    "/pipeline/task/{task_id}": {
      "get": {
        "description": "Return async task status by task ID.",
        "operationId": "get_pipeline_task_task_id",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineTaskStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Pipeline Task Status",
        "tags": [
          "Pipeline"
        ]
      }
    },
...
    "/recovery/events": {
      "get": {
        "description": "Get event timeline for recovery inspection.",
        "operationId": "get_recovery_events",
        "parameters": [
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "in": "query",
            "name": "job_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Job Id"
            }
          },
          {
            "in": "query",
            "name": "time_filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Time Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventTimelinePayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Recovery Events",
        "tags": [
          "Recovery"
        ]
      }
    },
...
    "/recovery/recommendations": {
      "get": {
        "description": "Get recovery recommendations for a job.",
        "operationId": "get_recovery_recommendations",
        "parameters": [
          {
            "in": "query",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationsResponsePayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Recovery Recommendations",
        "tags": [
          "Recovery"
        ]
      }
    },
    "/resources/association": {
      "post": {
        "description": "Build associative writing guidance from project state, retrieval evidence, and mode.",
        "operationId": "post_resources_association",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildAssociationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WritingAssociationPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Writing Association",
        "tags": [
          "Resources"
        ]
      }
    },
...
    "/resources/documents": {
      "get": {
        "description": "Get all document contents for a project (for RAG context).",
        "operationId": "get_resources_documents",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": {
                      "type": "string"
                    },
                    "type": "object"
                  },
                  "title": "Response Get Resources Documents",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project Documents",
        "tags": [
          "Resources"
        ]
      }
    },
...
    "/resources/drafts": {
      "get": {
        "description": "List all drafts, optionally filtered by section.",
        "operationId": "get_resources_drafts",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "section_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Section Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DraftPayload"
                  },
                  "title": "Response Get Resources Drafts",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Drafts",
        "tags": [
          "Resources"
        ]
      }
    },
...
    "/resources/materials": {
      "get": {
        "description": "List all materials attached to a project.",
        "operationId": "get_resources_materials",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MaterialPayload"
                  },
                  "title": "Response Get Resources Materials",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Materials",
        "tags": [
          "Resources"
        ]
      }
    },
...
  },
...
}